refactor(conv): move status-bar and queue rendering out of message.go
message.go had grown to fuse four concerns. status_bar.go's own header
already claimed to own the status bar, yet the orchestration
(OperationModeParams, RenderModeStatus, renderStatusCluster,
RenderTurnUsageSummary, compactStatusHint, the mode/thinking indicators)
lived in message.go next to its own dependencies (statusSegment,
RenderContextLabel, pctCritical). Move it into status_bar.go, and split
the unrelated queue preview into queue_preview.go.

message.go drops from 753 to 498 lines and is now just message + tool-
call rendering; status_bar.go holds the whole status line. No behaviour
change — same package, pure code movement.
