+++ internal/app/conv/message.go
@@ -10,8 +10,6 @@ import (
 
 	"github.com/genai-io/san/internal/app/kit"
 	"github.com/genai-io/san/internal/core"
-	"github.com/genai-io/san/internal/llm"
-	"github.com/genai-io/san/internal/setting"
 	"github.com/genai-io/san/internal/tool"
 )
 
@@ -31,162 +29,6 @@ const (
 	// beyond toolResultExpandedStyle's PaddingLeft(4). Total indent = 4 + 4 = 8 chars.
 	agentContentIndent = "    "
 )
-
-// OperationModeParams holds the parameters needed for rendering mode status.
-type OperationModeParams struct {
-	Mode             setting.OperationMode
-	InputTokens      int
-	InputLimit       int
-	ModelName        string
-	StatusMessage    string
-	ConversationCost llm.Money
-	Compressions     int  // session compact count, drives the "compacted ×N" badge
-	ShowContextBar   bool // render the visual [██████░░░░] 71% bar (opt-in)
-	Width            int
-	ThinkingEffort   string
-	ShowThinking     bool
-	QueueCount       int
-}
-
-// RenderModeStatus renders the combined mode status line.
-func RenderModeStatus(params OperationModeParams) string {
-	var leftParts []string
-
-	if modeStatus := RenderOperationModeIndicator(params.Mode); modeStatus != "" {
-		leftParts = append(leftParts, modeStatus)
-	}
-
-	if params.ShowThinking {
-		if thinkingStatus := RenderThinkingIndicator(params.ThinkingEffort); thinkingStatus != "" {
-			leftParts = append(leftParts, thinkingStatus)
-		}
-	}
-
-	if queueBadge := renderQueueBadge(params.QueueCount); queueBadge != "" {
-		leftParts = append(leftParts, queueBadge)
-	}
-
-	left := strings.Join(leftParts, "  ")
-
-	right := renderStatusCluster(params)
-	if right == "" || params.Width <= 0 {
-		return left
-	}
-
-	gap := max(2, params.Width-lipgloss.Width(left)-lipgloss.Width(right)-1)
-	return left + strings.Repeat(" ", gap) + right
-}
-
-// renderStatusCluster composes the status line's right-hand cluster, in
-// display order: model name, optional transient status message, the numeric
-// "ctx X/Y" label, the optional visual context bar, the optional compressions
-// badge, and the optional cost. Each piece is a statusSegment with a drop
-// priority; fitStatusSegments drops the least important first when the
-// terminal is too narrow to hold them all.
-func renderStatusCluster(p OperationModeParams) string {
-	if p.ModelName == "" {
-		return ""
-	}
-	muted := lipgloss.NewStyle().Foreground(kit.CurrentTheme.Muted)
-	sep := muted.Render(" · ")
-
-	// Priority 1 = most important (dropped last). The model name always
-	// renders; everything else drops before it under width pressure.
-	segments := []statusSegment{
-		{text: muted.Render(p.ModelName), priority: 1},
-	}
-	if p.StatusMessage != "" {
-		segments = append(segments, statusSegment{text: muted.Render(p.StatusMessage), priority: 2})
-	}
-
-	// The numeric label always renders — it falls back to "ctx X/--" when the
-	// limit is unknown, so the slot stays visible instead of silently hiding.
-	segments = append(segments, statusSegment{
-		text:     RenderContextLabel(p.InputTokens, p.InputLimit),
-		priority: 3,
-	})
-
-	// The visual bar is opt-in (off by default). When shown it also carries
-	// the auto-compact hint as a near-full warning.
-	if p.ShowContextBar {
-		bar := RenderContextBar(p.InputTokens, p.InputLimit)
-		if p.InputLimit > 0 {
-			if hint := compactStatusHint(float64(p.InputTokens) / float64(p.InputLimit) * 100); hint != "" {
-				bar += sep + muted.Render(hint)
-			}
-		}
-		segments = append(segments, statusSegment{text: bar, priority: 4})
-	}
-
-	if badge := RenderCompressionsBadge(p.Compressions); badge != "" {
-		segments = append(segments, statusSegment{text: badge, priority: 5})
-	}
-	if !p.ConversationCost.IsZero() {
-		segments = append(segments, statusSegment{text: muted.Render(kit.FormatMoney(p.ConversationCost)), priority: 6})
-	}
-
-	survivors := fitStatusSegments(segments, p.Width, lipgloss.Width(sep))
-	return strings.Join(survivors, sep)
-}
-
-func RenderTurnUsageSummary(inputTokens, outputTokens, width int) string {
-	if inputTokens == 0 && outputTokens == 0 {
-		return ""
-	}
-
-	summary := lipgloss.NewStyle().Foreground(kit.CurrentTheme.Muted).Render(
-		fmt.Sprintf("↑%s ↓%s", kit.FormatTokenCount(inputTokens), kit.FormatTokenCount(outputTokens)),
-	)
-	if width <= 0 {
-		return summary
-	}
-
-	gap := max(0, width-lipgloss.Width(summary))
-	return strings.Repeat(" ", gap) + summary
-}
-
-func compactStatusHint(percent float64) string {
-	switch {
-	case percent >= pctCritical:
-		return "auto-compact"
-	case percent > pctWarn:
-		return fmt.Sprintf("compact at %d%%", int(pctCritical))
-	default:
-		return ""
-	}
-}
-
-// RenderOperationModeIndicator returns the mode status indicator for auto-accept or bypass mode.
-func RenderOperationModeIndicator(mode setting.OperationMode) string {
-	var icon, label string
-	var clr kit.AdaptiveColor
-
-	switch mode {
-	case setting.ModeAutoAccept:
-		icon = "⏵⏵"
-		label = " accept edits on"
-		clr = kit.CurrentTheme.Success
-	case setting.ModeBypassPermissions:
-		icon = "⏵⏵"
-		label = " bypass permissions on"
-		clr = kit.CurrentTheme.Error
-	default:
-		return ""
-	}
-
-	style := lipgloss.NewStyle().Foreground(clr)
-	hint := lipgloss.NewStyle().Foreground(kit.CurrentTheme.Muted).Render(" (shift+tab to cycle)")
-	return "  " + style.Render(icon+label) + hint
-}
-
-func RenderThinkingIndicator(effort string) string {
-	if effort == "" || effort == "off" || effort == "none" {
-		return ""
-	}
-	style := lipgloss.NewStyle().Foreground(kit.CurrentTheme.Muted)
-	hint := lipgloss.NewStyle().Foreground(kit.CurrentTheme.Muted).Render(" (ctrl+t to cycle)")
-	return "  " + style.Render("✦ "+effort) + hint
-}
 
 // toolResultIcon returns the icon for tool results based on error state.
 func toolResultIcon(isError bool) string {
@@ -654,100 +496,3 @@ func stripMarkdownHeading(line string) string {
 	return indent + stripped
 }
 
-// QueuePreviewItem is the minimal data needed to render a queue item preview.
-type QueuePreviewItem struct {
-	Content   string
-	HasImages bool
-}
-
-var (
-	queueBadgeStyle = lipgloss.NewStyle().
-			Foreground(kit.CurrentTheme.Accent).
-			Bold(true)
-
-	queueContentStyle = lipgloss.NewStyle().
-				Foreground(kit.CurrentTheme.TextDim)
-
-	queueSelectedContentStyle = queueContentStyle.Foreground(kit.CurrentTheme.TextBright)
-
-	queueSelectedBadgeStyle = lipgloss.NewStyle().
-				Foreground(kit.CurrentTheme.TextBright).
-				Bold(true)
-
-	queueOverflowStyle = lipgloss.NewStyle().
-				Foreground(kit.CurrentTheme.Muted).
-				Italic(true)
-)
-
-// RenderQueuePreview renders queued input items above the input area.
-// selectedIdx is the currently selected item index (-1 = none).
-func RenderQueuePreview(items []QueuePreviewItem, selectedIdx, width int) string {
-	if len(items) == 0 {
-		return ""
-	}
-
-	var sb strings.Builder
-
-	maxVisible := 5
-	startIdx := 0
-	if len(items) > maxVisible && selectedIdx >= maxVisible {
-		startIdx = selectedIdx - maxVisible + 1
-	}
-	endIdx := min(startIdx+maxVisible, len(items))
-
-	for i := startIdx; i < endIdx; i++ {
-		item := items[i]
-		isSelected := i == selectedIdx
-
-		content := truncateQueueContent(item.Content, width-8)
-		if item.HasImages {
-			content = PendingImageStyle.Render("[Image] ") + content
-		}
-
-		if isSelected {
-			bar := kit.FocusBarStyle().Render(kit.FocusBar)
-			num := queueSelectedBadgeStyle.Render(fmt.Sprintf("%d.", i+1))
-			preview := queueSelectedContentStyle.Render(content)
-			fmt.Fprintf(&sb, " %s %s %s\n", bar, num, preview)
-		} else {
-			badge := queueBadgeStyle.Render(fmt.Sprintf("  %d.", i+1))
-			preview := queueContentStyle.Render(content)
-			fmt.Fprintf(&sb, " %s %s\n", badge, preview)
-		}
-	}
-
-	if len(items) > maxVisible {
-		if endIdx < len(items) {
-			sb.WriteString(queueOverflowStyle.Render(fmt.Sprintf("     +%d more below", len(items)-endIdx)) + "\n")
-		}
-		if startIdx > 0 {
-			above := queueOverflowStyle.Render(fmt.Sprintf("     +%d more above", startIdx)) + "\n"
-			return above + sb.String()
-		}
-	}
-
-	return sb.String()
-}
-
-// renderQueueBadge renders a compact badge for the status bar.
-func renderQueueBadge(count int) string {
-	if count == 0 {
-		return ""
-	}
-	return queueBadgeStyle.Render(fmt.Sprintf(" [%d queued]", count))
-}
-
-func truncateQueueContent(s string, maxLen int) string {
-	s = strings.ReplaceAll(s, "\n", " ")
-	s = strings.Join(strings.Fields(s), " ")
-
-	if maxLen <= 0 {
-		maxLen = 40
-	}
-
-	runes := []rune(s)
-	if len(runes) <= maxLen {
-		return s
-	}
-	return string(runes[:maxLen-1]) + "…"
-}
